requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
python-telegram-bot==20.7
playwright==1.41.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
cachetools==5.3.2
aiohttp==3.9.3
cssselect==1.2.0
Jinja2==3.1.3
python-multipart==0.0.6
beautifulsoup4==4.12.3
lxml==5.1.0
